# react/public/openapi3_1.yaml
openapi: "3.1.0"
info:
  version: "1.0.0"
  title: "Event ticketing app"
  description: "Project of a web application for selling event tickets using blockchain."
servers:
  - url: http://localhost:{port}/api
    description: Development server
    variables:
      port:
        enum:
          - "3000"
        default: "3000"
  - url: https://jalowiecki.toadres.pl/api
    description: Production server
paths:
...
  /users:
    get:
      operationId: ListUsers
      description: List available users
      security:
        - bearerAuth: [ADMINISTRATOR]
      parameters:
        - name: search
          description: A string to search user by username
          in: query
          schema:
            type: string
          required: false
        - name: page
          description: A listing page number
          in: query
          schema:
            $ref: "#/components/schemas/Page"
          required: false
      responses:
        "200":
          description: Available users
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    minimum: 0
                  pages:
                    type: integer
                    minimum: 1
                  currentPage:
                    type: integer
                    minimum: 1
                  results:
                    $ref: "#/components/schemas/ListUsers"
                required:
                  - count
                  - pages
                  - currentPage
                  - results
                additionalProperties: false
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Users
  /users/nonce:
    post:
      operationId: GenerateNonce
      description: Generate nonce for user with provided wallet address
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                publicAddress:
                  $ref: "#/components/schemas/PublicAddress"
              required:
                - publicAddress
              additionalProperties: false
      responses:
        "200":
          description: User nonce
          content:
            application/json:
              schema:
                type: object
                properties:
                  nonce:
                    $ref: "#/components/schemas/Nonce"
                required:
                  - nonce
                additionalProperties: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: User not found, return sign up token
          content:
            application/json:
              schema:
                type: object
                properties:
                  signUpToken:
                    $ref: "#/components/schemas/JWT"
                required:
                  - signUpToken
                additionalProperties: false
        "429":
          $ref: "#/components/responses/TooManyRequests"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Users
  /users/sign-up:
    post:
      operationId: SignUp
      description: Sign up user account
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                signUpToken:
                  $ref: "#/components/schemas/JWT"
                email:
                  $ref: "#/components/schemas/Email"
                username:
                  $ref: "#/components/schemas/Name"
                name:
                  $ref: "#/components/schemas/Name"
                surname:
                  $ref: "#/components/schemas/Name"
                birthdate:
                  $ref: "#/components/schemas/Date"
              required:
                - signUpToken
                - email
                - username
                - name
                - surname
                - birthdate
              additionalProperties: false
      responses:
        "204":
          description: Successfully signed up user account
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Users
  /users/activate:
    post:
      operationId: ActivateUser
      description: Activate user account
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                activationToken:
                  $ref: "#/components/schemas/JWT"
              required:
                - activationToken
              additionalProperties: false
      responses:
        "204":
          description: Successfully activated user account
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Users
  /users/verify:
    post:
      operationId: SignIn
      description: Sign in user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  minLength: 1
                signature:
                  description: ECDSA private key signature of message
                  type: string
                  minLength: 1
              required:
                - message
                - signature
              additionalProperties: false
      responses:
        "200":
          description: JWT tokens pair
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticationToken:
                    $ref: "#/components/schemas/JWT"
                  refreshToken:
                    $ref: "#/components/schemas/JWT"
                required:
                  - authenticationToken
                  - refreshToken
                additionalProperties: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Users
  /users/refresh:
    post:
      operationId: RefreshTokens
      description: Refresh JWT tokens pair
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  $ref: "#/components/schemas/JWT"
              required:
                - refreshToken
              additionalProperties: false
      responses:
        "200":
          description: JWT tokens pair
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticationToken:
                    $ref: "#/components/schemas/JWT"
                  refreshToken:
                    $ref: "#/components/schemas/JWT"
                required:
                  - authenticationToken
                  - refreshToken
                additionalProperties: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Users
  /users/{id}:
    parameters:
      - name: id
        description: The unique identifier of the user
        in: path
        schema:
          $ref: "#/components/schemas/Id"
        required: true
    patch:
      operationId: UpdateUserRole
      description: Update user role
      security:
        - bearerAuth: [ADMINISTATOR]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRole"
      responses:
        "200":
          description: Updated user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Users
  /users/me/events:
    get:
      operationId: ListUserEvents
      description: List events created by currently signed in user
      security:
        - bearerAuth: [EVENTS_ORGANIZER]
      parameters:
        - name: search
          description: A string to search event by name, tags and description
          in: query
          schema:
            type: string
          required: false
        - name: idSubcategory
          description: The unique identifier of the subcategory
          in: query
          schema:
            $ref: "#/components/schemas/Id"
          required: false
        - name: idCity
          description: The unique identifier of the city
          in: query
          schema:
            $ref: "#/components/schemas/Id"
          required: false
        - name: page
          description: A listing page number
          in: query
          schema:
            $ref: "#/components/schemas/Page"
          required: false
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    minimum: 0
                  pages:
                    type: integer
                    minimum: 1
                  currentPage:
                    type: integer
                    minimum: 1
                  results:
                    $ref: "#/components/schemas/ListEvents"
                required:
                  - count
                  - pages
                  - currentPage
                  - results
                additionalProperties: false
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Users
  /users/me/events/tickets:
    get:
      operationId: ListUserEventsTickets
      description: List tickets from events created by currently signed in user
      security:
        - bearerAuth: [EVENTS_ORGANIZER]
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                maxItems: 100
                $ref: "#/components/schemas/ListTickets"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Users
  /users/me/sales-report:
    get:
      operationId: GetSalesReport
      description: Get currently signed in user sales report details
      security:
        - bearerAuth: [EVENTS_ORGANIZER]
      responses:
        "200":
          description: Sales report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Users
...
components:
  responses:
...
    Unauthorized:
      description: Lack of or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: User is not permitted for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Page:
      description: A listing page number
      type: integer
      minimum: 1
...
    ListUsers:
      type: array
      items:
        type: object
        properties:
          idUser:
            $ref: "#/components/schemas/Id"
          publicAddress:
            $ref: "#/components/schemas/PublicAddress"
          email:
            $ref: "#/components/schemas/Email"
          username:
            $ref: "#/components/schemas/Name"
          role:
            $ref: "#/components/schemas/UserRole"
          active:
            type: boolean
          created:
            $ref: "#/components/schemas/DateTime"
        required:
          - idUser
          - publicAddress
          - email
          - username
          - role
          - active
          - created
        additionalProperties: false
...
  securitySchemes:
    bearerAuth:
      description: Bearer token using a JWT
      type: http
      scheme: bearer
      bearerFormat: JWT